Inline getCaretPosition helper
